Added washrooms to database

+++ backend/src/index.ts
@@ -7,7 +7,6 @@ const app = express();
 const PORT = 4000;
 let mongoURL: string;
 if (process.env.ENV === 'Docker') {
-    console.log("Docker");
     mongoURL = 'mongodb://mongodb:27017';
 } else {
     mongoURL = 'mongodb://127.0.0.1:27017';
@@ -15,6 +14,14 @@ if (process.env.ENV === 'Docker') {
 const dbName = "gohere"
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
+import { COLLECTIONS, Washroom, WashroomLocationReqPayload, isValidDatabaseWashroom } from "./databasetypes";
+import haversineDistance from 'haversine-distance';
+import { randomUUID } from 'crypto';
+import { jwtDecode } from 'jwt-decode';
+
+interface UsernameToken {
+    username: string
+}
 
 let db: Db;
 
@@ -36,10 +43,6 @@ app.listen(PORT, () => {
     console.log(`Server is running on http://localhost:${PORT}`);
   });
 
-const COLLECTIONS = {
-    users: "users",
-};
-
 app.post("/registerUser", express.json(), async (req, res) => {
     try {
         const { username, password } = req.body;
@@ -47,9 +50,7 @@ app.post("/registerUser", express.json(), async (req, res) => {
             return res.status(400).json({error: "Missing username or password."});
         }
 
-        console.log("Regieter");
-
-        const userCollection = db.collection(COLLECTIONS.users);
+        const userCollection = db.collection(COLLECTIONS.Users);
         const existingUser = await userCollection.findOne({ username });
         if (existingUser) {
             return res.status(400).json({ error: "Username already exists."})
@@ -80,12 +81,12 @@ app.post("/loginUser", express.json(), async (req, res) => {
         }
     
         // Find username in database
-        const userCollection = db.collection(COLLECTIONS.users);
+        const userCollection = db.collection(COLLECTIONS.Users);
         const user = await userCollection.findOne({ username });
     
         // Validate user against hashed password in database
         if (user && (await bcrypt.compare(password, user.password))) {
-            const token = jwt.sign({ username }, "secret-key", { expiresIn: "1h" });
+            const token = jwt.sign({ username: username }, "secret-key", { expiresIn: "1h" });
         
             // Send JSON Web Token to valid user
             res.json({ response: "User logged in succesfully.", token: token }); //Implicitly status 200
@@ -112,7 +113,7 @@ app.delete("/deleteUser", express.json(), async (req, res) => {
         return res.status(400).json({ error: "Missing username or password." });
         }
 
-        const userCollection = db.collection(COLLECTIONS.users);
+        const userCollection = db.collection(COLLECTIONS.Users);
         const user = await userCollection.findOne({ username });
     
         if (!(user && (await bcrypt.compare(password, user.password)))) {
@@ -120,8 +121,7 @@ app.delete("/deleteUser", express.json(), async (req, res) => {
         }
 
         jwt.verify(token, "secret-key", async (err, decoded) => {
-            if (err) { return res.status(401).send("Unauthorized."); }
-            console.log(decoded);
+            if (err || jwtDecode<UsernameToken>(token).username != username) { return res.status(401).send("Unauthorized."); }
             const result = await userCollection.deleteOne({ username: username });
             // Delete all other information
             
@@ -135,5 +135,91 @@ app.delete("/deleteUser", express.json(), async (req, res) => {
         res.status(500).json({ error: error.message });
     }
 });
+
+app.get("/getWashroomByLocation/:payload", express.json(), async (req, res) => {
+    const default_radius = 3000;
+    try {
+        const payload: string[] = req.params.payload.split('&');
+        const request: WashroomLocationReqPayload = {
+            latitude: parseFloat(payload[0]),
+            longitude: parseFloat(payload[1]),
+            radius: parseInt(payload[2]),
+        };
+        if (!request.latitude || !request.longitude) {
+            res.status(400).json( {error: "Longitude and latitude missing"} );
+            return;
+        }
+        if (!request.radius) {
+            request.radius = default_radius;
+        }
+        const userCollection = db.collection(COLLECTIONS.Washrooms);
+        const washrooms = (await userCollection.find().toArray()).filter((washroom) => {
+            // if (!washroom.approved) return false;
+            let dist = haversineDistance({ lat: washroom.latitude, lon: washroom.longitude },
+                              { lat: request.latitude, lon: request.longitude });
+            return dist <= request.radius!;
+        });
+
+        res.status(200).json({ response: washrooms })
+        
+    } catch (error: any) {
+        res.status(500).json({ error: error.message });
+    }
+});
+
+app.post("/addWashroom", express.json(), async (req, res) => {
+    try {
+
+        let new_washroom: Washroom = req.body;
+        new_washroom.approved = false;
+        new_washroom.id = randomUUID();
+        if (!isValidDatabaseWashroom(req.body)) {
+            res.status(404).json({ error: "Invalid request body." });
+            return;
+        }
+        if (req.headers.authorization) {
+            const token = req.headers.authorization.split(" ")[1];
+            console.log(token);
+            new_washroom.owner_username = jwtDecode<UsernameToken>(token).username;
+        }
+        const washroomCollection = db.collection(COLLECTIONS.Washrooms);
+        const result = washroomCollection.insertOne(new_washroom);
+        res.status(201).json({
+            response: "Washroom added successfully.",
+            washroomID: new_washroom.id,
+        })
+    } catch (error: any) {
+        res.status(500).json({ error: error.message });
+    }
+})
+
+app.delete("/deleteWashroom/:id", express.json(), async (req, res) => {
+    try {
+
+        let id = req.params.id;
+
+        if (!req.headers.authorization) {
+            res.status(401).json("Not Authorized.");
+            return;
+        }
+        const token = req.headers.authorization.split(" ")[1];
+
+        const washroomCollection = db.collection(COLLECTIONS.Washrooms);
+
+        jwt.verify(token, "secret-key", async (err, decoded) => {
+            if (err) { return res.status(401).send("Unauthorized."); }
+            const jwtDecoded = jwtDecode<UsernameToken>(token);
+            const result = await washroomCollection.deleteOne({owner_username: jwtDecoded.username, id: id});
+            
+            if (result.deletedCount === 1) {
+                res.status(200).json({ response: "Washroom with id " + id + " deleted."});
+            } else {
+                res.status(404).json({ error: "Could not find washroom associated with token to delete." })
+            }
+        });
+    } catch (error: any) {
+        res.status(500).json({ error: error.message });
+    }
+})
